Use asdf node version 16.15.0
asdf [1] is a tool for running multiple versions of different languages.
To use it one installs asdf itself, and then 'plugins'. For this case,
we will need the nodejs plugin [2].

Then, to use asdf, one runs:

```
asdf install nodejs 16.15.0
asdf reshim nodejs
```

The reshim is important, otherwise asdf won't pick up the latest
versions that have been installed for a given plugin.

To set a default version for a project one uses a .tool-versions file.

In this commit this file has been set to default nodejs to 16.15.0.

[1]: https://asdf-vm.com/
[2]: https://github.com/asdf-vm/asdf-nodejs