Bump to node 18.8.0
It appears that there is a bug in the 16.x.x series [1] that amongst
other things leads to memory leaks when attaching a memory watch process
to a noop process (i.e., () => {}).

To work around this, bump to the latest node.  18.x.x will soon be
active anyway [2], so this seems okay to do.

[1]: https://github.com/nodejs/node/issues/28787
[2]: https://nodejs.org/en/about/releases/